API Reference

Genomic matrices

BnGStructs.Haplotype — Type

struct Haplotype <: AbstractMatrix{Bool}

A struct to hold haplotype data in a compact BitMatrix form.

BnGStructs.Haplotype — Method

Haplotype(gt::BitMatrix)

Create a Haplotype struct from a BitMatrix gt. The number of loci and haplotypes are inferred from the size of gt. Only the valid part of gt (defined by nlc and nhp) is copied.

BnGStructs.Haplotype — Method

Haplotype(nlc::Int, nhp::Int)

Create a Haplotype struct with nlc loci and nhp haplotypes. The corresponding BitMatrix is initialized to all zeros.

BnGStructs.Genotype — Type

struct Genotype <: AbstractMatrix{Bool}

A struct to store SNP genotypes in a BitMatrix, with individuals as rows and alleles as columns.

BnGStructs.Genotype — Method

Genotype(gt::BitMatrix)

Create a Genotype struct from a BitMatrix gt. The number of individuals and alleles are inferred from the size of gt. Only the valid part of gt (defined by nid and nas) is copied.

BnGStructs.Genotype — Method

Genotype(nid::Int, nas::Int)

Create a Genotype struct with nid individuals and nas alleles. The corresponding BitMatrix is initialized to all zeros.

BnGStructs.hap2id — Method

hap2id(hps::Haplotype)

This is to transform a Haplotype object to a Genotype object.

Given the SNP haplotypes hps of type Haplotype, this function transposes its gt field (of size nlc × 2nid) to a BitMatrix of size n × 2nlc, where n is the minimum number that is a multiple of 64 and is no less than nid. The resulting BitMatrix is used to construct and return a Genotype object.

Note this function deems the computer is of 64-bit architecture.

BnGStructs.id2hap — Method

id2hap(g::Genotype)

This is to transform a Genotype object back to a Haplotype object, performing the reverse operation of hap2id.

Given a Genotype object g, this function transposes its gt field (of size nid × 2nlc) to a BitMatrix of size m × 2nid, where m is the minimum number that is a multiple of 64 and is no less than nlc. The resulting BitMatrix is used to construct and return a Haplotype object.

Note this function deems the computer is of 64-bit architecture.

Species

BnGStructs.Cat — Type

struct Cat <: Species
    name::String
    nid::Int32
    chromosome::Vector{UInt32}
    M::UInt32
end

Struct for cat species.

BnGStructs.Cat — Method

Cat(nid::Integer; M::Integer = 100_000_000)

This function creates a Cat object with the given nid. The chromosome lengths are from https://www.ncbi.nlm.nih.gov/datasets/genome/GCF_018350175.1/

BnGStructs.Cattle — Type

struct Cattle <: Species
    name::String
    nid::Int32
    chromosome::Vector{UInt32} # Changed to UInt32
    M::UInt32 # Changed to UInt32
end

Struct for cattle species.

BnGStructs.Cattle — Method

Cattle(nid::Integer; M::Integer = 100_000_000)

Helper function to create a cattle species struct. The chromosome lengths are from https://www.ncbi.nlm.nih.gov/datasets/genome/GCF_000003055.6/

BnGStructs.Chicken — Type

struct Chicken <: Species
    name::String
    nid::Int32
    chromosome::Vector{UInt32} # Changed to UInt32
    M::UInt32 # Changed to UInt32
end

Struct for chicken species.

BnGStructs.Chicken — Method

Chicken(nid::Integer; M::Integer = 100_000_000)

This function creates a Chicken object with the given nid. The chromosome lengths are from https://www.ncbi.nlm.nih.gov/datasets/genome/GCF_016699485.2/

BnGStructs.Dog — Type

struct Dog <: Species
    name::String
    nid::Int32
    chromosome::Vector{UInt32}
    M::UInt32
end

Struct for dog species.

BnGStructs.Dog — Method

Dog(nid::Integer; M::Integer = 100_000_000)

This function creates a Dog object with the given nid. The chromosome lengths are from https://www.ncbi.nlm.nih.gov/datasets/genome/GCF_011100685.1/

BnGStructs.GenericSpecies — Type

struct GenericSpecies <: Species
    name::String
    nid::Int32
    chromosome::Vector{UInt32}
    M::UInt32
end

Struct for a generic species with the given name, population size, and chromosome lengths.

BnGStructs.GenericSpecies — Method

GenericSpecies(name::AbstractString, nid::Integer, chromosome::AbstractVector{<:Integer}; M::Integer=100_000_000)

Construct a GenericSpecies with name, population size nid, chromosome lengths chromosome (in bp), and base pairs per Morgan M.

BnGStructs.Goat — Type

struct Goat <: Species
    name::String
    nid::Int32
    chromosome::Vector{UInt32}
    M::UInt32
end

Struct for goat species.

BnGStructs.Goat — Method

Goat(nid::Integer; M::Integer = 100_000_000)

This function creates a Goat object with the given nid. The chromosome lengths are from https://www.ncbi.nlm.nih.gov/datasets/genome/GCF_001704415.2/

BnGStructs.Horse — Type

struct Horse <: Species
    name::String
    nid::Int32
    chromosome::Vector{UInt32}
    M::UInt32
end

Struct for horse species.

BnGStructs.Horse — Method

Horse(nid::Integer; M::Integer = 100_000_000)

This function creates a Horse object with the given nid. The chromosome lengths are from https://www.ncbi.nlm.nih.gov/datasets/genome/GCF_002863925.1/

BnGStructs.Pig — Type

struct Pig <: Species
    name::String
    nid::Int32
    chromosome::Vector{UInt32} # Changed to UInt32
    M::UInt32 # Changed to UInt32
end

Struct for pig species.

BnGStructs.Pig — Method

Pig(nid::Integer; M::Integer = 100_000_000)

This function creates a Pig object with the given nid. The chromosome lengths are from https://www.ncbi.nlm.nih.gov/datasets/genome/GCF_000003025.6/

BnGStructs.Rabbit — Type

struct Rabbit <: Species
    name::String
    nid::Int32
    chromosome::Vector{UInt32}
    M::UInt32
end

Struct for rabbit species.

BnGStructs.Rabbit — Method

Rabbit(nid::Integer; M::Integer = 100_000_000)

This function creates a Rabbit object with the given nid. The chromosome lengths are from https://www.ncbi.nlm.nih.gov/datasets/genome/GCF_000003625.3/

BnGStructs.Sheep — Type

struct Sheep <: Species
    name::String
    nid::Int32
    chromosome::Vector{UInt32} # Changed to UInt32
    M::UInt32 # Changed to UInt32
end

Struct for sheep species.

BnGStructs.Sheep — Method

Sheep(nid::Integer; M::Integer = 100_000_000)

This function creates a Sheep object with the given nid. The chromosome lengths are from https://www.ncbi.nlm.nih.gov/datasets/genome/GCF_000298735.2/

BnGStructs.Species — Type

abstract type Species end

BnGStructs.M — Method

M(sp::Species)

Return the number of base pairs per Morgan.

BnGStructs.cbp — Method

cbp(sp::Species)

Return cumulative chromosome endpoints in base pairs.

BnGStructs.chromosome — Method

chromosome(sp::Species)

Return chromosome lengths in base pairs.

BnGStructs.name — Method

name(sp::Species)

Return the species name.

BnGStructs.nid — Method

nid(sp::Species)

Return the population size.

BnGStructs.total_bp — Method

total_bp(sp::Species)

Return the total autosomal genome length in base pairs.

Traits and SNP sets

BnGStructs.AbstractTrait — Type

AbstractTrait

Abstract type for all traits.

BnGStructs.Trait — Type

abstract type Trait <: AbstractTrait end # non-threshold traits

Non-threshold traits have a field μ.

BnGStructs.Trait — Method

Trait(
    name::AbstractString,
    weight::AbstractVector{<:Real};
    sex = 2,
    age = 1.0,
    h² = 0.25,
    QTL = :qtl,
    σₐ = 1.0,
    da = Normal(),
)

Helper function to create a threshold tTrait. It checks the validity of the input parameters and returns a tTrait. Refer to the tTrait struct for the description of the parameters. The weight, of minimum length 2, defines the relative ratio of the phenotype categories. The length of the weight is one more than the length of the thresholds in struct tTrait. This function then decides the thresholds based on the weight and standard normal distribution of the genotype.

The underlying TBV distribution is Normal(0, 1.0). The categorical phenotypes are calculated from the phenotypic thresholds calculated from Weights.

BnGStructs.Trait — Method

Trait(
    name;
    sex = 2,
    age = 0.,
    h² = 0.25,
    QTL = :qtl,
    μ = 0.0,
    σₐ = 1.0,
    da = Normal(),
)

Helper function to create a Trait object. It checks the validity of the input parameters and returns a Trait object. Refer to the Trait struct for the description of the parameters.

BnGStructs.aTrait — Type

struct aTrait{D<:Distribution} <: Trait
    name::String # must be a valid Julia identifier
    sex::Int     # expressed in ♀(0), or ♂(1), or 2 for both sexes.
    age::Float64 # age that trait is measured
    h²::Float64  # narrow-sense heritability
    QTL::Symbol  # QTL column name in lmp
    μ::Float64   # population mean
    σₐ::Float64  # TBV std
    da::D        # QTL additive effect distribution, e.g., Normal()
end

Pure additive trait.

BnGStructs.adTrait — Type

struct adTrait <: Trait
    name::AbstractString
end

Trait with additive and dominance effects. To be implemented later.

BnGStructs.adiTrait — Type

struct adiTrait <: Trait
    name::String
end

Trait with additive, dominance, and epistatic effects. To be implemented later.

BnGStructs.ditTrait — Type

struct ditTrait <: AbstractTrait
    name::String
end

Threshold trait with dominance, and epistatic effects. To be implemented later.

BnGStructs.dtTrait — Type

struct dtTrait <: AbstractTrait
    name::String
end

Threshold trait with dominance effects. To be implemented later.

BnGStructs.tTrait — Type

struct tTrait{D<:Distribution} <: AbstractTrait
    name::String # must be a valid Julia identifier
    threshold::Vector{Float64} # phenotype will be 0, 1, ...
    sex::Int     # expressed in ♀(0), or ♂(1), 2 for both sexes.
    age::Float64 # age that trait is measured
    h²::Float64  # narrow-sense heritability
    QTL::Symbol
    σₐ::Float64  # TBV std
    da::D        # e.g., Normal()
end

Struct for threshold traits. The number of phenotype categories equals the length of threshold + 1. The underlying genotype is continuous. The phenotype will be one of 0, 1, ..., length(threshold). Population mean is not used for threshold traits.

BnGStructs.SNPSet — Type

struct SNPSet
    name::String
    nlc::Int
    maf::Float64
    exclusive::Bool
end

BnGStructs.SNPSet — Method

SNPSet(name::String, nlc::Int; maf = 0.0, exclusive = false)

Create a SNPSet object with the given name, number of loci, minor allele frequency (MAF), and exclusivity flag.

Base.show — Method

Base.show(io::IO, snp::SNPSet)

Display the SNPSet object snp in a human-readable format.

Variant maps and QTL architectures

BnGStructs.VariantMap — Type

struct VariantMap

A compact struct-of-arrays representation of genomic variant metadata.

Fields

chr::Vector{Int8}: Chromosome numbers.

pos::Vector{UInt32}: Base-pair coordinates on the chromosome.

ref::Vector{Char}: Reference alleles (e.g. 'A', 'C', 'G', 'T').

alt::Vector{Char}: Alternate alleles.

frq::Vector{Float32}: Alternate allele frequencies (optional, may be empty).

BnGStructs.VariantMap — Method

VariantMap(chr::Vector{<:Integer}, pos::Vector{<:Integer}, ref::Vector{Char}, alt::Vector{Char}; frq = Float32[])

Outer constructor converting integer vectors to standard Int8 and UInt32.

BnGStructs.LocusSet — Type

struct LocusSet

A named subset of variant loci identified by their 1-based indices into a VariantMap.

Commonly used to define:

Visible SNP chip panels (e.g. 50k / 770k marker sets)

Reference / background tracking loci

Candidate causal regions

BnGStructs.MultiTraitQTL — Type

struct MultiTraitQTL

Joint multi-trait QTL architecture containing a shared set of unique QTL loci and a (K × T) effect matrix across T traits.

Enables fast BLAS matrix multiplications for calculating True Breeding Values (TBV) across multiple traits simultaneously.

BnGStructs.MultiTraitQTL — Method

MultiTraitQTL(traits::Vector{TraitQTL})

Construct a unified MultiTraitQTL by merging multiple single-trait TraitQTL objects, pooling all unique QTL loci into a shared set.

BnGStructs.TraitQTL — Type

struct TraitQTL

Sparse representation of QTL loci and their effect sizes for a single trait.

Fields

name::String: Trait name.

loci::Vector{Int32}: 1-based indices of QTL loci into a VariantMap.

additive::Vector{Float64}: Additive effect sizes (α) per QTL locus.

dominance::Vector{Float64}: Optional dominance effect sizes (d) per QTL locus (empty if purely additive).

BnGStructs.tbv — Method

tbv(gt::Genotype, mqtl::MultiTraitQTL) -> Matrix{Float64}

Calculate True Breeding Values (TBV) for all individuals across multiple traits (nid × n_traits) from a Genotype and a MultiTraitQTL.

BnGStructs.tbv — Method

tbv(gt::Genotype, qtl::TraitQTL) -> Vector{Float64}

Calculate True Breeding Values (TBV) for all individuals from a Genotype (nid × 2nlc) and a TraitQTL.

BnGStructs.tbv — Method

tbv(hap::Haplotype, mqtl::MultiTraitQTL) -> Matrix{Float64}

Calculate True Breeding Values (TBV) for all individuals across multiple traits (nid × n_traits) from a Haplotype and a MultiTraitQTL.

BnGStructs.tbv — Method

tbv(hap::Haplotype, qtl::TraitQTL) -> Vector{Float64}

Calculate True Breeding Values (TBV) for all individuals from a Haplotype (nlc × 2nid) and a TraitQTL.